+++ src/types/DecimalTuple.ts
@@ -1,0 +1,6 @@
+/**
+ * 2-Tuple to represent decimals [coefficient, exponent]
+ */
+export type DecimalTuple = readonly [number, number];
+
+export default DecimalTuple;